Create test:integ command with integ/ folder
Move all handlers / repositories over there. Can move them back gradually in debugging the open handles.
