+++ package.json
@@ -11,7 +11,8 @@
     "build": "tsc",
     "clean": "rm -rf dist cdk.out",
     "watch": "tsc -w",
-    "test": "jest --detectOpenHandles --forceExit --testPathPattern=test/unit/",
+    "test": "jest --testPathPattern=test/unit/",
+    "test:integ": "jest --detectOpenHandles --forceExit --testPathPattern=test/integ/",
     "test:e2e": "jest --detectOpenHandles --runInBand --testPathPattern=test/e2e/",
     "cdk": "cdk",
     "fix": "run-s fix:*",
